Timer Driver development
The aim of the driver is to generate an event periodically after a specified amount of microseconds. Here are the driver specifications:

- It should have a start function that starts the periodic event generation (initializes too) with two arguments: a integer specifying the time in us and a function pointer specifying the callback function.

- It should have a stop function stops the periodic event generation.

Use the driver in a simple application that prints a counter value that increments every 509 ms. As you don't have this microcontroller with you, you can just write the code and send it to us. For anything not specified here, you can assume specs yourself and state it explicitly.